requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-ollama==0.1.0
langchain-huggingface==0.0.3
langchain-text-splitters>=0.2.4
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
scikit-learn>=1.3.0
emoji>=2.0.0
rank-bm25==0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
ragas==0.1.4
gigachat==0.2.0